refactor: refactor error page files and components
