+++ packages/author-profile/src/author-profile-list-error-prop-types.js
@@ -1,0 +1,7 @@
+import PropTypes from "prop-types";
+
+const propTypes = {
+  refetch: PropTypes.func.isRequired
+};
+
+export default propTypes;
